// src/renderer/src/enums/formatKind.ts
/** The formatters, by name; the value is the `kind` attribute's word. */
export enum FormatKind {
  Number = 'number',
  Whole = 'whole',
  Signed = 'signed',
  Gold = 'gold',
  Percent = 'percent',
  Duration = 'duration',
  Minutes = 'minutes',
  Played = 'played',
  Date = 'date',
  Time = 'time',
  Datetime = 'datetime',
  When = 'when',
  Relative = 'relative',
  Until = 'until'
}